Remove the "feedback" feature and unused feedback panel
