--- app/assets/stylesheets/components/page_specific/_feedback_panel.scss
@@ -1,4 +1,0 @@
-.feedback-panel__title {
-  @extend %heading-extra-small;
-  margin-top:0;
-}
