+++ app/views/styleguide/pages_guide.html.erb
@@ -8,19 +8,6 @@
       <%= render 'styleguide/shared/guide' %>
 
       <%= render 'shared/related_links', categories: categories %>
-
-      <section class="panel feedback-panel">
-        <h2 class="feedback-panel__title">Give us your feedback</h2>
-
-        <p>What is your feedback about?</p>
-
-        <ul>
-          <li><a href="/en/styleguide/pages/feedback_information?no-nav">The information on this page</a></li>
-          <li><a href="/en/styleguide/pages/feedback_technical?no-nav">I had a technical issue</a></li>
-          <li><a href="/en/styleguide/pages/feedback_advice?no-nav">I need money advice</a></li>
-        </ul>
-
-      </section>
     </main>
   </div>
 </div>
