+++ config/features.yml.sample
@@ -1,7 +1,6 @@
 features:
   page_feedback: true
   budget_planner_social_share: true
-  feedback: false
   optimizely: false
   cms_home_page: true
   cms_home_page_promo_banner: false
