+++ config/locales/cy.yml
@@ -509,13 +509,6 @@ cy:
       saturday_times_html: "dydd Sadwrn: <strong>9am–1pm</strong>"
       about_call_charges: Dysgwch am gostau galwadau
 
-  feedback_panel:
-    title: Rhowch eich adborth i ni
-    greeting: Beth yw natur eich adborth?
-    article_feedback_link_text: Y wybodaeth ar y dudalen hon
-    technical_feedback_link_text: Roedd gennyf fater technegol
-    advice_link_text: Mae arnaf angen cymorth ariannol
-
   inline_feedback:
     technical_feedback_link: Adrodd ar anhawster technegol gyda'r dudalen hon
