+++ config/locales/en.yml
@@ -330,7 +330,7 @@ en:
       user_feedback_submit: Submit
       feedback_placeholder: Type here...
       empty_validation: Please leave a comment
-      
+
     sharing:
       title: Share this article
       label_facebook: Share this article on Facebook
@@ -505,13 +505,6 @@ en:
     show:
       title: Help and money advice
       greeting: "For help and money advice, please contact us:"
-
-  feedback_panel:
-    title: Give us your feedback
-    greeting: What is your feedback about?
-    article_feedback_link_text: The information on this page
-    technical_feedback_link_text: I had a technical issue
-    advice_link_text: I need money advice
 
   inline_feedback:
     technical_feedback_link: Report a technical issue with this page
