+++ features/support/ui/pages/article.rb
@@ -1,6 +1,5 @@
 require_relative '../page'
 require_relative '../sections/category_nav'
-require_relative '../sections/feedback_panel'
 
 module UI::Pages
   class Article < UI::Page
@@ -10,8 +9,8 @@ module UI::Pages
     element :related_categories, '.related-categories'
     element :breadcrumbs, '.l-context-bar'
     element :end_of_article_signup, '.js-end-of-article-signup'
+    element :technical_feedback_link, '.t-inline-feedback__technical-feedback-link'
 
     section :category_nav, UI::Sections::CategoryNav, 'nav .link-list-primary'
-    section :feedback_panel, UI::Sections::FeedbackPanel, '.feedback-panel'
   end
 end
